split up code into helper function
