+++ src/task/methods/third_place/config.vsh.yaml
@@ -38,10 +38,11 @@ functionality:
     - type: python_script
       path: script.py
     - path: submission(14).csv
+    - path: helper.py
 
 platforms:
   - type: docker
-    image: ghcr.io/openproblems-bio/base_pytorch_nvidia:1.0.4
+    image: ghcr.io/openproblems-bio/base_python:1.0.4
     setup:
       - type: python
         packages: [ tensorflow==2.12.0, numpy==1.23.5, pandas==2.0.3, matplotlib==3.7.2, scikit-learn==1.3.2, pyarrow==13.0.0, fastparquet ]
